+++ lib/src/analytics/analytics_wrapper.dart
@@ -1,15 +1,15 @@
 import 'package:flutter/cupertino.dart';
 import 'package:loglytics/loglytics.dart';
 
-/// Used to wrap implementations of [AnalyticsSubjects] and [AnalyticsParameters] for a specific feature.
+/// Used to wrap implementations of [LoglyticsSubjects] and [LoglyticsParameters] for a specific feature.
 ///
-/// Each feature in your app should have its own [AnalyticsWrapper] implementation with its own
-/// respective [AnalyticsSubjects] and [AnalyticsParameters] implementations. Check out the [Loglytics]
+/// Each feature in your app should have its own [LoglyticsWrapper] implementation with its own
+/// respective [LoglyticsSubjects] and [LoglyticsParameters] implementations. Check out the [Loglytics]
 /// documentation for more info on how to use these as part of your analytics solution.
 ///
 /// Template:
 ///
-/// class TemplateAnalytics extends AnalyticsWrapper<TemplateSubjects, TemplateParameters> {
+/// class TemplateAnalytics extends LoglyticsWrapper<TemplateSubjects, TemplateParameters> {
 ///   @override
 ///   TemplateSubjects get subjects => _counterSubjects;
 ///   late final TemplateSubjects _counterSubjects = TemplateSubjects();
@@ -19,19 +19,19 @@ import 'package:loglytics/loglytics.dart';
 ///   late final TemplateParameters _counterParameters = TemplateParameters();
 /// }
 ///
-/// class TemplateSubjects extends AnalyticsSubjects {}
+/// class TemplateSubjects extends LoglyticsSubjects {}
 ///
-/// class TemplateParameters extends AnalyticsParameters {}
+/// class TemplateParameters extends LoglyticsParameters {}
 @immutable
-abstract class AnalyticsWrapper<S extends AnalyticsSubjects, P extends AnalyticsParameters> {
+abstract class LoglyticsWrapper<S extends LoglyticsSubjects, P extends LoglyticsParameters> {
   S get subjects;
   P get parameters;
 }
 
 /// Used to specify analytic subject names.
 @immutable
-abstract class AnalyticsSubjects {}
+abstract class LoglyticsSubjects {}
 
 /// Used to specify analytic parameter names.
 @immutable
-abstract class AnalyticsParameters {}
+abstract class LoglyticsParameters {}