+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 * **✨ New:** Added default analytics that are accessible trough `Loglytics.defaultAnalytics` and require no further configuration of the `Loglytics` mixin.
 * **⚠️ Breaking:** Renamed the following:
-  * `FeatureAnalytics` to `AnalyticsWrapper`;
-  * `FeatureSubjects` to `AnalyticsSubjects`;
-  * `FeatureParameters` to `AnalyticsParameters`;
+  * `FeatureAnalytics` to `LoglyticsWrapper`;
+  * `FeatureSubjects` to `LoglyticsSubjects`;
+  * `FeatureParameters` to `LoglyticsParameters`;
   * `Loglytics.featureAnalytics` getter to `Loglytics.wrapper`.
 * **✨ New:** Added extra event types and methods.
 * **⚠️ Breaking:** Removed create, update, delete event types and methods (these are better fit as subjects).